Removed dev dependency on Jupyter
- split away notebook testing in a separate file
- ignored that file
- added a note in examples/_notebooks
- fixed a typing error with cli
- updated deps
