+++ examples/_notebooks/README.md
@@ -1,0 +1,3 @@
+# Important note
+
+Running vsketch in notebooks, although possible, is no longer supported. This example may or may not run, and will at least require explicit installation of jupyter (`pip install jupyter`).
