+++ pyproject.toml
@@ -43,10 +43,6 @@ pytest = ">=6.2.4"
 pytest-benchmark = "^3.4.1"
 mypy = ">=0.901"
 types-requests = ">=0.1.9"
-jupyter = "^1.0.0"
-nbconvert = "^6.0.7"
-nbformat = "^5.0.8"
-jupyter-client = ">=6.1.7"
 isort = ">=5.10.1"
 black = ">=22.1"
 
@@ -56,6 +52,7 @@ requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.pytest.ini_options]
+addopts = "--ignore=tests/test_notebooks.py"  # no longer supported
 testpaths = ["tests"]
 
 [tool.black]
