+++ tests/test_examples.py
@@ -1,25 +1,13 @@
 import pathlib
 
-import nbformat
 import pytest
 from click.testing import CliRunner
-from nbconvert.preprocessors import ExecutePreprocessor
 
 from vsketch_cli.cli import cli
 
 EXAMPLES = pathlib.Path(__file__).parent.parent.absolute() / "examples"
-NOTEBOOKS = EXAMPLES / "_notebooks"
 
 runner = CliRunner()
-
-
-@pytest.mark.slow
-@pytest.mark.parametrize("ipynb", NOTEBOOKS.glob("*.ipynb"))
-def test_example_notebooks(tmp_path, ipynb):
-    with open(ipynb) as f:
-        nb = nbformat.read(f, as_version=4)
-    ep = ExecutePreprocessor(timeout=600, kernel_name="python3")
-    ep.preprocess(nb, {"metadata": {"path": tmp_path}})
 
 
 @pytest.mark.parametrize("path", EXAMPLES.glob("[!._]*[!.md]"))
